// src/app/easy/models/CustomerUBS.ts
export class CustomerUBS {
  business_date: Date;
  cat_id: string;
  segment_code: string;
  branch_code: string;
  cust_no: string;
  customer_name: string;
  customer_type: string;
  birth_date: Date;
  cif_creation_date: Date;
  casa_balance: number;
  fd_balance: number;
  total_balance: number;
  priority_code: string;
  family_cif: string;
  status: string;
  address: string;
  mobile_number: string;
  e_mail: string;
  birthday_recom: string;
  picdrop_recom: string;
  birthday_reason: string;
  picdrop_reason: string;
  branch_name: string;
  cat_desc : string;
  segment_name :string;
  customer_desc : string;
}