// src/app/easy/models/ServiceRequest.ts
export class ServiceRequest {
    work_date: Date;
    service_slno: string;
    service_code: string;
    branch_code: string;
    priority_code: string;
    cust_no: string;
    cust_name: string;
    cust_address: string;
    cust_mobile: string;
    cust_email: string;
    cust_birthday: Date;
    casa_bal: number;
    fd_bal: number;
    tot_bal: number;
    alternet_name: string;
    alternet_mobile: string;
    alternet_email: string;
    airline_code: string;
    airline_name: string;
    flight_no: string;
    reporting_date: string;
    airline_code_return: string;
    airline_name_return: string;
    flight_no_return: string;
    reporting_date_return: string;
    total_person: string;
    vendor_code: string;
    vehichle_type: string;
    pick_address: string;
    drop_address: string;
    return_airport: string;
    drop_airport: string;
    gift_code: string;
    cust_feedback: string;
    remarks: string;
    recom_note: string;
    apv_note: string;
    status_code: string;
    ins_by: string;
    ins_date: Date;
    upd_by: string;
    upd_date: Date;
    recom_by: string;
    recom_date: Date;
    apv_by: string;
    apv_date: Date;
    userid: string;
    total_person_return: string;
    vehichle_type_return: string;
    vendor_name: string;
    gift_name: string;
    service_name: string;
    recom_message: string;
    segment_code: string;
    branch_name: string;
}